+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fixed image url in the README.
+- Fixed animation display when the first animation is not just Data.
 
 ## [0.6.0] - 2022-03-17